Move entity package to repository package
