+++ backend/spring-app/spring-app/src/main/java/com/avdo/spring/app/controller/ProductController.java
@@ -1,13 +1,11 @@
 package com.avdo.spring.app.controller;
 
 import com.avdo.spring.app.controller.dto.CreateProductDto;
-import com.avdo.spring.app.entity.ProductEntity;
 import com.avdo.spring.app.service.ProductService;
 import com.avdo.spring.app.service.domain.model.Product;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.support.DefaultMessageSourceResolvable;
-import org.springframework.http.ResponseEntity;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
