+++ backend/spring-app/spring-app/src/main/java/com/avdo/spring/app/repository/OrderItemRepository.java
@@ -1,9 +1,7 @@
 package com.avdo.spring.app.repository;
 
-import com.avdo.spring.app.entity.OrderItemEntity;
+import com.avdo.spring.app.repository.entity.OrderItemEntity;
 import com.avdo.spring.app.service.domain.model.OrderItem;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
 public interface OrderItemRepository {
