+++ backend/spring-app/spring-app/src/main/java/com/avdo/spring/app/repository/crud/CrudCartItemRepository.java
@@ -1,6 +1,6 @@
 package com.avdo.spring.app.repository.crud;
 
-import com.avdo.spring.app.entity.CartItemEntity;
+import com.avdo.spring.app.repository.entity.CartItemEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.transaction.annotation.Transactional;
