+++ backend/spring-app/spring-app/src/main/java/com/avdo/spring/app/repository/crud/CrudCartRepository.java
@@ -1,6 +1,6 @@
 package com.avdo.spring.app.repository.crud;
 
-import com.avdo.spring.app.entity.CartEntity;
+import com.avdo.spring.app.repository.entity.CartEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
