+++ backend/spring-app/spring-app/src/main/java/com/avdo/spring/app/repository/crud/CrudOrderItemRepository.java
@@ -1,6 +1,6 @@
 package com.avdo.spring.app.repository.crud;
 
-import com.avdo.spring.app.entity.OrderItemEntity;
+import com.avdo.spring.app.repository.entity.OrderItemEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CrudOrderItemRepository extends JpaRepository<OrderItemEntity, Long> {
