+++ backend/spring-app/spring-app/src/main/java/com/avdo/spring/app/repository/crud/CrudOrderRepository.java
@@ -1,6 +1,6 @@
 package com.avdo.spring.app.repository.crud;
 
-import com.avdo.spring.app.entity.OrderEntity;
+import com.avdo.spring.app.repository.entity.OrderEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CrudOrderRepository extends JpaRepository<OrderEntity, Long> {
