+++ backend/spring-app/spring-app/src/main/java/com/avdo/spring/app/repository/entity/ProductEntity.java
@@ -1,4 +1,4 @@
-package com.avdo.spring.app.entity;
+package com.avdo.spring.app.repository.entity;
 
 import com.avdo.spring.app.service.domain.model.Product;
 import com.fasterxml.jackson.annotation.JsonIdentityInfo;
