+++ backend/spring-app/spring-app/src/main/java/com/avdo/spring/app/repository/impl/CategoryJpaRepository.java
@@ -1,6 +1,6 @@
 package com.avdo.spring.app.repository.impl;
 
-import com.avdo.spring.app.entity.CategoryEntity;
+import com.avdo.spring.app.repository.entity.CategoryEntity;
 import com.avdo.spring.app.repository.CategoryRepository;
 import com.avdo.spring.app.repository.crud.CrudCategoryRepository;
 import com.avdo.spring.app.service.domain.model.Category;
