+++ backend/spring-app/spring-app/src/main/java/com/avdo/spring/app/repository/impl/OrderItemJpaRepository.java
@@ -1,6 +1,6 @@
 package com.avdo.spring.app.repository.impl;
 
-import com.avdo.spring.app.entity.OrderItemEntity;
+import com.avdo.spring.app.repository.entity.OrderItemEntity;
 import com.avdo.spring.app.repository.OrderItemRepository;
 import com.avdo.spring.app.repository.crud.CrudOrderItemRepository;
 import com.avdo.spring.app.service.domain.model.OrderItem;
