+++ backend/spring-app/spring-app/src/main/java/com/avdo/spring/app/repository/impl/OrderJpaRepository.java
@@ -5,6 +5,7 @@ import com.avdo.spring.app.repository.CartRepository;
 import com.avdo.spring.app.repository.OrderItemRepository;
 import com.avdo.spring.app.repository.OrderRepository;
 import com.avdo.spring.app.repository.crud.CrudOrderRepository;
+import com.avdo.spring.app.repository.entity.*;
 import com.avdo.spring.app.service.domain.model.Cart;
 import com.avdo.spring.app.service.domain.model.Order;
 import com.avdo.spring.app.service.domain.model.OrderItem;
