+++ backend/spring-app/spring-app/src/main/java/com/avdo/spring/app/repository/impl/ProductJpaRepository.java
@@ -1,15 +1,13 @@
 package com.avdo.spring.app.repository.impl;
 
-import com.avdo.spring.app.entity.CategoryEntity;
-import com.avdo.spring.app.entity.ProductEntity;
+import com.avdo.spring.app.repository.entity.CategoryEntity;
+import com.avdo.spring.app.repository.entity.ProductEntity;
 import com.avdo.spring.app.repository.ProductRepository;
 import com.avdo.spring.app.repository.crud.CrudProductRepository;
 import com.avdo.spring.app.service.CategoryService;
-import com.avdo.spring.app.service.ProductService;
 import com.avdo.spring.app.service.domain.model.Category;
 import com.avdo.spring.app.service.domain.model.Product;
 import com.avdo.spring.app.service.domain.request.CreateProductRequest;
-import com.avdo.spring.app.service.impl.ProductServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
