+++ backend/spring-app/spring-app/src/main/java/com/avdo/spring/app/service/OrderItemService.java
@@ -1,10 +1,7 @@
 package com.avdo.spring.app.service;
 
-import com.avdo.spring.app.entity.OrderItemEntity;
-import com.avdo.spring.app.repository.OrderItemRepository;
+import com.avdo.spring.app.repository.entity.OrderItemEntity;
 import com.avdo.spring.app.service.domain.model.OrderItem;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
 
 public interface OrderItemService {
