+++ backend/spring-app/spring-app/src/main/java/com/avdo/spring/app/service/domain/model/Product.java
@@ -1,6 +1,6 @@
 package com.avdo.spring.app.service.domain.model;
 
-import com.avdo.spring.app.entity.OrderItemEntity;
+import com.avdo.spring.app.repository.entity.OrderItemEntity;
 import lombok.Builder;
 import lombok.Getter;
