+++ backend/spring-app/spring-app/src/main/java/com/avdo/spring/app/service/domain/model/User.java
@@ -1,6 +1,6 @@
 package com.avdo.spring.app.service.domain.model;
 
-import com.avdo.spring.app.entity.OrderEntity;
+import com.avdo.spring.app.repository.entity.OrderEntity;
 import lombok.Builder;
 import lombok.Getter;
