+++ backend/spring-app/spring-app/src/main/java/com/avdo/spring/app/service/impl/OrderItemServiceImpl.java
@@ -1,6 +1,6 @@
 package com.avdo.spring.app.service.impl;
 
-import com.avdo.spring.app.entity.OrderItemEntity;
+import com.avdo.spring.app.repository.entity.OrderItemEntity;
 import com.avdo.spring.app.repository.OrderItemRepository;
 import com.avdo.spring.app.service.OrderItemService;
 import com.avdo.spring.app.service.domain.model.OrderItem;
